feat: Extract the `DetailForm` components
