+++ src/pages/Editor/GGEditor/components/EditorDetailPanel/MindDetailPanel.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import { Card } from 'antd';
 import { NodePanel, CanvasPanel, DetailPanel } from 'gg-editor';
-import NodeDetail from './NodeDetail';
+import DetailForm from './DetailForm';
 import styles from './index.less';
 
 const MindDetailPanel = () => {
   return (
     <DetailPanel className={styles.detailPanel}>
       <NodePanel>
-        <NodeDetail />
+        <DetailForm type="node" />
       </NodePanel>
       <CanvasPanel>
         <Card type="inner" size="small" title="Canvas" bordered={false} />
